services/trade_producer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["shubhangiranjan2011 <shubhangiranjan2011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
quixstreams = "^2.5.1"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
